@@ -18,7 +18,7 @@
 		"@babel/core": "7.14.3",
 		"@babel/preset-env": "7.14.2",
 		"@babel/preset-react": "7.13.13",
-		"autoprefixer": "10.2.5",
+		"autoprefixer": "10.2.6",
 		"babel-loader": "8.2.2",
 		"chalk": "4.1.1",
 		"commander": "7.2.0",
@@ -27,7 +27,7 @@
 		"cssnano": "5.0.4",
 		"extract-loader": "5.1.0",
 		"extract-text-webpack-plugin": "3.0.2",
-		"ink": "^2.7.1",
+		"ink": "3.0.8",
 		"json-loader": "0.5.7",
 		"lodash": "4.17.21",
 		"node-emoji": "1.10.0",
@@ -36,7 +36,7 @@
 		"postcss-loader": "5.3.0",
 		"prop-types": "^15.7.2",
 		"raw-loader": "4.0.2",
-		"react": "^16.14.0",
+		"react": "^17.0.2",
 		"sass": "1.34.0",
 		"sass-loader": "11.1.1",
 		"strip-ansi": "7.0.0",
@@ -51,5 +51,8 @@
 		"webpack-cli": "4.7.0",
 		"webpack-merge": "5.7.3",
 		"webpack-sources": "2.2.0"
+	},
+	"devDependencies": {
+		"@types/react": "17.0.8"
 	}
 }
