@@ -1,28 +1,11 @@
-import { Box, Text } from "ink";
 import React from "react";
-import FeedbackTemplate from "../../helpers/FeedbackTemplate";
-import useCommand from "../../helpers/useCommand";
-const { spawn } = require("child_process");
-
-const path = require("path");
+import CompilationFeedback from "../../helpers/CompilationFeedback";
+import webpackConfig from "../../webpack/scripts/webpack.dev.js";
 
 function WatchScripts() {
-	const webpackConfigPath = path.resolve(
-		process.cwd(),
-		`node_modules/shopify-builder/config/scripts/webpack.dev.js`
+	return (
+		<CompilationFeedback webpackConfig={webpackConfig} type="scripts" watch />
 	);
-
-	const command = [
-		`cross-env NODE_ENV=development`,
-		` ${path.resolve(process.cwd(), "node_modules/.bin/webpack")}`,
-		`--progress --watch`,
-		`--config="${webpackConfigPath}"`,
-	].join(" ");
-
-	const { output, error } = useCommand(command);
-	console.log("output", output);
-	console.log("error", error);
-	return <FeedbackTemplate output={output} error={error} />;
 }
 
 export default WatchScripts;
