@@ -1,23 +1,11 @@
 import React from "react";
-import FeedbackTemplate from "../../helpers/FeedbackTemplate";
-import useCommand from "../../helpers/useCommand";
-const path = require("path");
+import CompilationFeedback from "../../helpers/CompilationFeedback";
+import webpackConfig from "../../webpack/sections/webpack.dev.js";
 
 function WatchSections() {
-	const webpackConfigPath = path.resolve(
-		process.cwd(),
-		`node_modules/shopify-builder/config/sections/webpack.dev.js`
+	return (
+		<CompilationFeedback webpackConfig={webpackConfig} type="sections" watch />
 	);
-	const command = [
-		`cross-env NODE_ENV=production`,
-		` ${path.resolve(process.cwd(), "node_modules/.bin/webpack")}`,
-		`--progress --watch`,
-		`--config="${webpackConfigPath}"`,
-	].join(" ");
-
-	const { output, error } = useCommand(command);
-
-	return <FeedbackTemplate output={output} error={error} />;
 }
 
 export default WatchSections;
